{"schema":"decodex/commit/1","summary":"Move scroll capture resume frontier","authority":"manual"}
